fix(website): Route mobile download CTAs to Get Started
