+++ CHANGELOG.md
@@ -67,6 +67,7 @@ This project uses selective package publishing. Each release entry lists the pub
 - The WebRTC transport is now actually functional: nodes load node-datachannel at startup (previously it was never initialized, so every connection silently used TCP). Peers with a working WebRTC stack advertise `transport.webrtc.v1` in discovery metadata. Encrypted TCP remains the preferred transport between nodes (WebRTC data channels cap messages at ~256 KiB, below the 1 MiB protocol chunk size); initiators use WebRTC only toward peers that support WebRTC but not encrypted TCP. Sellers without a working stack now refuse `hello` signaling cleanly instead of crashing on it — previously any WebRTC connection attempt (e.g. from a browser buyer) hit an unguarded code path.
 - Free sellers are no longer penalized for omitting cached-input pricing: a $0 offer has nothing to underquote, so it skips the 50% effective-reputation reduction. Under the hard minimum-reputation gate that reduction had made every such free seller ineligible for model-only auto routing whenever any paid competitor on the same model advertised cached pricing (a halved 0–100 score can never reach the default 60). The Desktop first-launch free model default now also requires a free seller that actually passes the auto-selection gate — a model that only looked free off a gated-out cheapest seller could previously become the default and silently dispatch a zero-balance user to a paid peer, failing with an out-of-credits error.
 - Auto seller selection now treats the minimum trust score purely as a gate and orders eligible peers by total price, with reputation, failure streaks, and preference bonuses breaking price ties. Previously reputation kept dominating the ranking above the threshold (one display point of reputation outweighed $2/Mtok of price), so a 9.9-reputation paid seller would beat an eligible free seller of the same model — Auto now picks the free or cheapest eligible offer, matching the Price + Trust preference the UI describes.
+- Website VPR call-to-actions no longer offer a desktop installer download on phones. The navbar item in the mobile hamburger menu and the two ANTS token page buttons now read **Get Started** and route to the `/get-started` mobile onboarding flow, matching the behavior the homepage buttons already had.
 
 ### Added
