+++ apps/website/src/theme/NavbarItem/DownloadNavbarItem.tsx
@@ -8,10 +8,17 @@ import {useLatestDesktopDownload} from '@site/src/lib/useLatestDesktopDownload';
  * runtime to the direct installer for the visitor's OS/arch, falling back to
  * the releases page when no installer matches (Linux, unknown platforms, or
  * before the release lookup resolves).
+ *
+ * In the mobile sidebar (hamburger menu) the item becomes "Get Started"
+ * linking to the /get-started Telegram flow instead — phones and tablets
+ * can't run the desktop installer, so a download link is a dead end there.
  */
 export default function DownloadNavbarItem(
   props: Omit<DefaultNavbarItemProps, 'href' | 'to'>,
 ): React.ReactNode {
   const {href} = useLatestDesktopDownload();
+  if (props.mobile) {
+    return <DefaultNavbarItem {...props} label="Get Started" to="/get-started" />;
+  }
   return <DefaultNavbarItem {...props} href={href} />;
 }
